requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-community
langchain-chroma
langchain-core
unstructured
selenium
chromadb
sentence-transformers
pandas
plotly
beautifulsoup4
requests
